20190822_03_ユーザ認証アプリ追加
