+++ accounts/admin.py
@@ -1,0 +1,5 @@
+'''
+管理者画面を表示するためのコード
+'''
+
+# Register your models here.
