+++ accounts/models.py
@@ -1,0 +1,3 @@
+'''
+テーブルの情報を定義するファイル
+'''
