+++ accounts/tests.py
@@ -1,0 +1,3 @@
+'''
+テスト用コード作成ファイル
+'''
