+++ config/urls.py
@@ -28,6 +28,7 @@ urlpatterns = [
     path('admin/', admin.site.urls),
     # URLの指定なしの場合、searchapp内のurls.pyに指定されているURLの処理に移動する
     path('', include('searchapp.urls')),
+    path('accounts/',include('accounts.urls')),
 ]
