+++ searchapp/forms.py
@@ -3,9 +3,6 @@
 
 '''
 from django import forms
-from django.contrib.auth.forms import (
-    AuthenticationForm
-)
 from .models import CategoryTBL
 
 
@@ -120,15 +117,3 @@ class ColorForm(forms.Form):
                 },
             ),
     )
-
-
-class LoginForm(AuthenticationForm):
-    '''ログインフォーム'''
-
-    def __init__(self, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-        for field in self.fields.values():
-            # 使用するclassを定義する。
-            field.widget.attrs['class'] = 'form-control'
-            # placeholder(入力欄に初期表示する内容)を定義する。
-            field.widget.attrs['placeholder'] = field.label
