+++ static/accounts/css/style.css
@@ -1,0 +1,4 @@
+
+h1 {
+    margin-bottom: 5px;
+}
