--- templates/registration/base.html
@@ -1,18 +1,0 @@
-{% load staticfiles %}
-<!DOCTYPE html>
-<html lang="ja">
-    <head>
-        <meta charset="utf-8">
-        <meta name="viewport" content="width=device-width, initial-scale=1">
-        <!-- bootstrap4の呼び出し -->
-        {% load bootstrap4 %}
-        {% bootstrap_css %}
-        {% bootstrap_javascript jquery='full' %}
-        <link href="{% static 'css/style.css' %}" rel="stylesheet">
-        <script type="text/javascript" src="{% static 'js/searchapp.js' %}"></script>
-        <title>{% block title %}ECサイト{% endblock %}</title>
-    </head>
-            <!-- オーバーライド対象：content=子テンプレートのブロック名 -->
-            {% block content %}
-            {% endblock %}
-</html>
