--- templates/searchapp/base.html
@@ -1,47 +1,0 @@
-{% load staticfiles %}
-<!DOCTYPE html>
-<html lang="ja">
-    <head>
-        <meta charset="utf-8">
-        <meta name="viewport" content="width=device-width, initial-scale=1">
-        <!-- bootstrap4の呼び出し -->
-        {% load bootstrap4 %}
-        {% bootstrap_css %}
-        {% bootstrap_javascript jquery='full' %}
-        <link href="{% static 'css/style.css' %}" rel="stylesheet">
-        <script type="text/javascript" src="{% static 'js/searchapp.js' %}"></script>
-        <title>{% block title %}ECサイト{% endblock %}</title>
-    </head>
-    <body>
-        <header class="site-header">
-            {% if user.is_authenticated %}
-                <div class="container">
-                    <span>ようこそ {{ user.get_username }}様</span>
-                    <a href="{% url 'searchapp:logout' %}?next={{ request.path }}">ログアウト</a>
-                </div>
-            {% else %}
-                <div class="container">
-                    <span>ようこそ ゲスト様</span>
-                    <a href="{% url 'searchapp:login' %}?next={{ request.path }}">ログイン</a>
-                </div>
-            {% endif %}
-        </header>
-        <main>
-            <!-- <h1 id="site-name"><a href="{% url 'admin:index' %}">管理者画面</a></h1>  -->
-
-            {% if messages %}
-                <div class="alert alert-warning alert-dismissible fade show" role="alert">
-                    <button type="button" class="close" data-dismiss="alert" aria-label="Close">
-                        <span aria-hidden="true">&times;</span>
-                    </button>
-                    {% for message in messages %}
-                        <p> {{ message }}</p>
-                    {% endfor %}
-                </div>
-            {% endif %}
-            <!-- オーバーライド対象：content=子テンプレートのブロック名 -->
-            {% block content %}
-            {% endblock %}
-        </main>
-    </body>
-</html>
