+++ templates/searchapp/details.html
@@ -1,4 +1,17 @@
-{% extends 'searchapp/base.html' %} {% block content %}
+<!-- テンプレートの継承先 -->
+{% extends 'base.html' %}
+
+{% block title %}商品詳細{% endblock %}
+
+{% block header %}
+    {% include 'accounts/authheader.html' %}
+{% endblock %}
+
+{% block msg %}
+    {% include 'accounts/logoutmsg.html' %}
+{% endblock %}
+
+{% block body %}
 <!-- humanizeの呼び出し -->
 {% load humanize %}
 <div class="container mar">
